feat: add popover to status column in events table
